back/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.10
et_xmlfile==2.0.0
exceptiongroup==1.2.0
executing==2.0.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
ipython==8.18.1
jedi==0.19.1
jiter==0.8.2
matplotlib-inline==0.1.6
npm==0.1.1
numpy==2.0.2
oauthlib==3.2.2
openai==1.59.7
openpyxl==3.1.5
optional-django==0.1.0
packaging==25.0
pandas==2.2.3
parso==0.8.3
pexpect==4.9.0
pillow==11.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.17.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.12
